Select all CCDs of the matching visits by default
Add an option and change the default behavior
to select all CCDs of the visits that overlap,
instead of selecting only CCDs that overlap.
They are needed in the mosaic/jointcal step.
For jointcal, exposures of all ccds on each
visit to build the full focal plane model are
needed, not just the ccds overlapping the tract.